Add clang-tidy integration: zero findings across 291 files, and chaos testing spec
- .clang-tidy: bugprone-*, selected modernize-*/performance-*/readability-*/
  cppcoreguidelines-*, clang-analyzer-*; WarningsAsErrors: "*"; HeaderFilterRegex
  scoped to project-owned headers
- CMakeLists.txt: static-analysis (parallel via run-clang-tidy) and
  static-analysis-fix targets; compile_commands.json guard
- scripts/pre-commit-coverage.sh: opt-in tidy step between format and coverage
  (TIDY_CHECK=1 to enable, SKIP_TIDY_CHECK=1 to skip)
- Codebase fixes to reach zero findings: narrowing conversions (static_cast),
  enum sizes (uint8_t), switch defaults, else-after-return, branch-clone,
  use-after-move NOLINT, compiler diagnostic errors in future.hpp and
  coap_transport_impl.hpp
- clang-format applied to all modified files
- README.md: Static Analysis section; doc/TODO.md updated
- .kiro/specs/clang-tidy/: requirements, design, tasks (all complete)
- .kiro/specs/chaos-testing/: new spec for libfiu fault injection chaos testing;
  macro-based approach (fiu_do_on compiles to no-op without FIU_ENABLE);
  21 tasks covering build integration, fault points, and 8 property tests

Skip coverage: executor_concept_requirements_property_test has a pre-existing
memory access violation unrelated to these changes (confirmed by running the
same test on the unmodified tree).

+++ tests/error_handler_exponential_backoff_unit_test.cpp
@@ -83,8 +83,8 @@ BOOST_AUTO_TEST_CASE(test_exponential_backoff_calculation, *boost::unit_test::ti
     for (std::size_t i = 1; i < test_max_attempts; ++i) {
         auto delay = test_initial_delay;
         for (std::size_t j = 1; j < i; ++j) {
-            delay = std::chrono::milliseconds{
-                static_cast<long long>(delay.count() * test_backoff_multiplier)};
+            delay = std::chrono::milliseconds{static_cast<long long>(
+                static_cast<double>(delay.count()) * test_backoff_multiplier)};
         }
         delay = std::min(delay, test_max_delay);
         expected_min_delay += delay;
